requirements.txt
flask
requests
PyNaCl
python-dotenv
